Refactor: 메인 페이지까지 주석 추가 완료
